LabProject14 bug fix
카메라 회전에 관한 버그 수정

+++ LabProjects/LabProject14/Player.h
@@ -59,8 +59,9 @@ public:
 	// xmf3Position 벡터에서 현재 플레이어의 위치 벡터를 빼면 현재 플레이어의 위치에서 
 	// xmf3Position 방향으로의 벡터가 된다. 현재 플레이어의 위치에서 이 벡터 만큼 이동한다.
 	void SetPosition(const XMFLOAT3& xmf3Position) {
-		Move(XMFLOAT3(xmf3Position.x -
-			m_xmf3Position.x, xmf3Position.y - m_xmf3Position.y, xmf3Position.z - m_xmf3Position.z),
+		Move(XMFLOAT3(xmf3Position.x - m_xmf3Position.x, 
+			xmf3Position.y - m_xmf3Position.y, 
+			xmf3Position.z - m_xmf3Position.z),
 			false);
 	}
 
@@ -73,9 +74,10 @@ public:
 	void SetCamera(CCamera* pCamera) { m_pCamera = pCamera; }
 
 	//플레이어를 이동하는 함수이다. 
-	void Move(ULONG nDirection, float fDistance, bool bVelocity = false);
+	void Move(DWORD nDirection, float fDistance, bool bVelocity = false);
 	void Move(const XMFLOAT3& xmf3Shift, bool bVelocity = false);
 	void Move(float fxOffset = 0.0f, float fyOffset = 0.0f, float fzOffset = 0.0f);
+
 	//플레이어를 회전하는 함수이다. 
 	void Rotate(float x, float y, float z);
